# Feature/effect freer monad optimizations