--- a/src/Application/NBB.Application.Effects/Mediator.cs
+++ b/src/Application/NBB.Application.Effects/Mediator.cs
@@ -5,6 +5,6 @@ namespace NBB.Application.Effects
 {
     public static class Mediator
     {
-        public static Effect<TResponse> SendQuery<TResponse>(IQuery<TResponse> query) => Effect.Of(new MediatorSendQuery.SideEffect<TResponse>(query));
+        public static Effect<TResponse> SendQuery<TResponse>(IQuery<TResponse> query) => Effect.Of<MediatorSendQuery.SideEffect<TResponse>, TResponse>(new MediatorSendQuery.SideEffect<TResponse>(query));
     }
 }