--- a/src/Core/NBB.Core.Effects.FSharp/Interpreter.fs
+++ b/src/Core/NBB.Core.Effects.FSharp/Interpreter.fs
@@ -4,12 +4,12 @@ open NBB.Core.Effects
 
 module Interpreter =
 
-    type SideEffectMediator() =
-        interface ISideEffectMediator with
+    type SideEffectBroker() =
+        interface ISideEffectBroker with
             member _.Run((sideEffect: ISideEffect<'a>), cancellationToken) =
                 match sideEffect with
                 | :? Thunk.SideEffect<'a> -> Thunk.Handler<'a>().Handle(sideEffect:?> Thunk.SideEffect<'a>, cancellationToken)
                 | _ -> failwith "Invalid sideEffect"
              
 
-    let createInterpreter = SideEffectMediator >> Interpreter
+    let createInterpreter = SideEffectBroker >> Interpreter