--- a/src/Core/NBB.Core.Effects.FSharp/Interpreter.fs
+++ b/src/Core/NBB.Core.Effects.FSharp/Interpreter.fs
@@ -1,15 +1,5 @@
 ﻿namespace NBB.Core.Effects.FSharp
 
-open NBB.Core.Effects
 
 module Interpreter =
-
-    type SideEffectBroker() =
-        interface ISideEffectBroker with
-            member _.Run((sideEffect: ISideEffect<'a>), cancellationToken) =
-                match sideEffect with
-                | :? Thunk.SideEffect<'a> -> Thunk.Handler<'a>().Handle(sideEffect:?> Thunk.SideEffect<'a>, cancellationToken)
-                | _ -> failwith "Invalid sideEffect"
-             
-
-    let createInterpreter = SideEffectBroker >> Interpreter
+    let createInterpreter = NBB.Core.Effects.Interpreter.CreateDefault