--- a/src/Core/NBB.Core.Effects/DependencyInjectionExtensions.cs
+++ b/src/Core/NBB.Core.Effects/DependencyInjectionExtensions.cs
@@ -11,6 +11,7 @@ namespace NBB.Core.Effects
         {
             services.AddSingleton(typeof(Thunk.Handler<>));
             services.AddSingleton(typeof(Parallel.Handler<,>));
+            services.AddSingleton(typeof(Sequenced.Handler<>));
             services.AddScoped<ISideEffectBroker, SideEffectBroker>();
             services.AddScoped<IInterpreter, Interpreter>();
         }
@@ -31,7 +32,7 @@ namespace NBB.Core.Effects
         public static IServiceCollection AddSideEffectHandler<TSideEffect, TOutput>(this IServiceCollection services, Func<TSideEffect, TOutput> handlerFn)
             where TSideEffect : ISideEffect<TOutput>
         {
-            Task<TOutput> HandlerFnAsync(TSideEffect sideEffect, CancellationToken _cancellationToken)
+            Task<TOutput> HandlerFnAsync(TSideEffect sideEffect, CancellationToken cancellationToken)
             {
                 return Task.FromResult(handlerFn(sideEffect));
             }