--- a/src/Core/NBB.Core.Effects/ISideEffect.cs
+++ b/src/Core/NBB.Core.Effects/ISideEffect.cs
@@ -1,7 +1,7 @@
 ﻿namespace NBB.Core.Effects
 {
 
-    public interface ISideEffect<out TOutput>
+    public interface ISideEffect<out TResult>
     {
     }