… src/Core/NBB.Core.Effects/Fun.cs
… src/Core/NBB.Core.Effects/IAmHandledBy.cs
… src/Core/NBB.Core.Effects/IEffectVisitor.cs
… src/Core/NBB.Core.Effects/ISideEffectBroker.cs
… src/Core/NBB.Core.Effects/ISideEffectHandler.cs
… src/Core/NBB.Core.Effects/Interpreter.cs
… src/Core/NBB.Core.Effects/IterativeInterpreterVisitor.cs
… src/Core/NBB.Core.Effects/NBB.Core.Effects.csproj
… src/Core/NBB.Core.Effects/ParallelSideEffect.cs
… src/Core/NBB.Core.Effects/SequencedSideEffect.cs
… src/Core/NBB.Core.Effects/SideEffectBroker.cs
… src/Core/NBB.Core.Effects/ThunkSideEffect.cs
… src/Core/NBB.Core.Effects/Unit.cs
… src/Http/NBB.Http.Effects/Http.cs
… src/Messaging/NBB.Messaging.Effects/DependencyInjectionExtensions.cs
… src/Messaging/NBB.Messaging.Effects/MessageBus.cs
… src/Messaging/NBB.Messaging.Effects/PublishMessage.cs
… src/Orchestration/NBB.ProcessManager.Definition/SideEffects/RequestTimeout.cs
… src/Orchestration/NBB.ProcessManager.Definition/SideEffects/Timeout.cs
… src/Orchestration/NBB.ProcessManager.Runtime/DependencyInjectionExtensions.cs
… src/Orchestration/NBB.ProcessManager.Runtime/SideEffectHandlers/CancelTimeoutsHandler.cs
… src/Orchestration/NBB.ProcessManager.Runtime/SideEffectHandlers/RequestTimeoutHandler.cs
… test/Benchmarks/EffectsBenchmarks/EffectsBenchmarks/EffectBenchmark.fs
… test/UnitTests/Core/NBB.Core.Effects.FSharp.Tests/Tests.fs
… test/UnitTests/Core/NBB.Core.Effects.Tests/EffectCompositionTests.cs
… test/UnitTests/Core/NBB.Core.Effects.Tests/InterpreterTests.cs
… test/UnitTests/Core/NBB.Core.Effects.Tests/SideEffectBrokerTests.cs
… test/UnitTests/Core/NBB.Core.Effects.Tests/ThunkSideEffectTests.cs
… test/UnitTests/Messaging/NBB.Messaging.Effects.Tests/MessagingEffectsTest.cs
… test/UnitTests/Orchestration/NBB.ProcessManager.Tests/InstanceFixture.cs