- Implemented Impure using a type alligned continuation list
 - Parallel SideEffect
 - implemented Apply using Parallel SideEffect
 - Sequenced SideEffect
 - Imlemented Sequence using Sequenced SideEffect
 - Impure has the type param SideEffect
 - SideEffectBroker optimizations
 - Removed ISideEffectHandler<T> and ISideEffectHandler

The Effect BEnchmark yields:
Method | N | Mean | Error | StdDev
RunLazy | 50 | 97.20 μs | 1.117 μs | 0.664 μs
RunStrict | 50 | 74.74 μs | 1.933 μs | 1.279 μs
RunLazy | 500 | 796.56 μs | 13.906 μs | 9.198 μs
RunStrict | 500 | 516.24 μs | 20.672 μs | 13.673 μs
RunLazy | 5000 | 13,532.42 μs | 832.443 μs | 550.609 μs
RunStrict | 5000 | 8,774.10 μs | 236.151 μs | 156.199 μs